### 2019-11-11 #1874 @johannes-titz
Sure!

@rmhogervorst, would you be willing to review this software paper for JOSS?

@russellpierce, would you be willing to review this software paper for JOSS?
