### 2019-11-18 #1874 @russellpierce
Sure, I'd be happy to review this.  Is there a link available for author's copies of the associated Titz & Sedlmeier, 2019 paper?
