### 2019-11-19 #1874 @johannes-titz
@russellpierce thank you! The paper is not published yet, but I will send you a copy of the submitted version right away via e-mail.
