### 2019-11-19 #1900 @whedon
[REVIEW]: passt: An R implementation of the Probability Associator Time (PASS-T) model
**Submitting author:** @johannes-titz (<a href="http://orcid.org/0000-0002-1102-5719">Johannes Titz</a>)
**Repository:** <a href="https://github.com/johannes-titz/passt" target ="_blank">https://github.com/johannes-titz/passt</a>
**Version:** v0.1.1
