### 2019-12-05 089c226
fix class problem with new devel version
class returns c("matrix", "array") for new versions of R:

class(matrix(1 : 4, 2, 2))
  [1] "matrix" "array"

In calc_output_sum I check for class, which had to be adapted.
