### 2019-12-05 857638c
use methods::is instead of class
