+++ R/passt.R
@@ -164,8 +164,9 @@ updt_winner_weights <- function(input, weight_matrix, lrate){
 #' @inheritParams updt_winner_weights
 #' @return sum of activations of output units for input patterns
 #' @noRd
+#' @importFrom methods is
 calc_output_sum <- function(inputs, weight_matrix){
-  if (class(inputs)[1] == "matrix") {
+  if (methods::is(inputs, "matrix")) {
     return(colSums(apply(inputs, 1, function(x) weight_matrix %*% x)))
   }
   if (is.vector(inputs) == TRUE) {
