+++ NAMESPACE
@@ -7,6 +7,7 @@ importFrom(dplyr,arrange)
 importFrom(dplyr,mutate)
 importFrom(dplyr,summarize)
 importFrom(magrittr,"%>%")
+importFrom(methods,is)
 importFrom(rlang,.data)
 importFrom(stats,cor)
 importFrom(stats,rnorm)
