+++ DESCRIPTION
@@ -19,7 +19,7 @@ Description: Simulates judgments of frequency and duration based on
 License: GPL-3
 Encoding: UTF-8
 LazyData: true
-RoxygenNote: 6.1.1
+RoxygenNote: 7.0.2
 URL: https://github.com/johannes-titz/passt
 BugReports: https://github.com/johannes-titz/passt/issues
 Suggests: 
@@ -29,4 +29,4 @@ Suggests:
     testthat (>= 2.1.0),
     covr
 VignetteBuilder: knitr
-Imports: magrittr,dplyr,tidyr,rlang
+Imports: magrittr,methods,dplyr,tidyr,rlang
