### 2019-12-05 35bada1
test both release and devel, warnings are errors
