+++ .travis.yml
@@ -4,3 +4,8 @@ language: R
 cache: packages
 after_success:
     - Rscript -e 'covr::codecov()'
+warnings_are_errors: true
+sudo: required
+r:
+  - release
+  - devel
