### 2019-12-05 65bca1f
increase version number and update news.md
