+++ NEWS.md
@@ -1,3 +1,8 @@
+# passt 0.1.1
+
+* fix problem in *calc_output_sum* which occurs because the behavior of the 
+function *class* changed in the new version of R
+
 # passt 0.1.0
 
 * First Release
