+++ NEWS.md
@@ -1,3 +1,5 @@
+#
+
 # passt 0.1.1
 
 * fix problem in *calc_output_sum* which occurs because the behavior of the 
