### 2019-12-06 #1900 @johannes-titz
Just for your information: I had to submit a patch to CRAN because the behavior of the *class* function has changed in the development version of R. So version 0.1.1 of my package is out now. If you happen to work with the development version or run tests on it, version 0.1.0 will throw an error. In the current version of R (3.6.1) passt v.0.1.0 still works fine. But note that version 3.6.2 is already in prerelease state and the final release is scheduled for 2019-12-12. If you cannot get the review done until then, I suggest that you review version 0.1.1. Not sure how this is best handled. If you want to install from CRAN, you will also get version 0.1.1 by default from now on. Basically, this is unavoidable since without the patch the package would be taken down from CRAN. @oliviaguest is it possible to change the version when the submission is still under review? I think in this specific case this would make sense.
