### 2019-12-20 #1900 @nicksexton
Hi @oliviaguest @johannes-titz  I'm happy to say I'm satisfied this meets JOSS guidelines. @johannes-titz Nice work on this, I have to say that while the instructions on how to use the package are good, it's not that easy to understand the cognitive detail of the model. I would be happier to see a more complete description of the cognitive model - ie. more formal than the instructions on how to use the package to run a simulation and get results, and higher level than just diving straight into the source code. However I assume these details are in the papers that are under review and I think is outside the scope of the JOSS review, so I'm happy to accept this.
