### 2019-12-24 #1900 @oliviaguest
@nicksexton thank you — hopefully the authors can link to their papers from their code repo.
