### 2019-12-27 #1900 @johannes-titz
Thank you for the feedback! In the papers that are under review the cognitive model is explained "verbally" and formally in detail. As soon as the papers are published, I will update the references (add a doi). For now, I do not want to upload a preprint version of the submitted papers since I stated that the material has not been disseminated before. The submitted papers already link to the code repository.
