### 2019-12-27 #1900 @oliviaguest
@johannes-titz FWIW and perhaps for future reference, you can check out what policy the journal you have submitted to has w.r.t. preprints: http://sherpa.ac.uk/romeo/index.php
