### 2020-01-11 #1900 @oliviaguest
Thanks, OK.

1. I propose we wait to see what @russellpierce means by two weeks: please elaborate if you meant that literally. And..

2. we wait to how @johannes-titz feels about having to wait probably another 3-4 weeks (assuming 2 weeks is a rough estimate but not too far off) before getting something back from @russellpierce.  Are you OK with this set-up and if not can you propose another reviewer, perhaps?
