### 2020-01-13 #1900 @johannes-titz
@oliviaguest @russellpierce I can wait another 3-4 weeks. Thanks for the update.

@oliviaguest Thanks a lot for the website to check the preprint policy of journals. I am still a bit hesitant as I might submit to other journals (with other policies) if the papers are rejected (which is always a risk). If you feel an explanation of the cognitive model is needed, I will add one in the documentation on github.
