### 2020-01-28 #1900 @russellpierce
Thank you very much for the ping.  I've been able to do a first pass review the software; I want to kick the functionality tires a bit more and dig into the software paper.  But, there is progress at least.  I have one more hour on my schedule for tomorrow after which I'll wrap things up one way or another.  So far, very good!
