### 2020-02-03 #1900 @oliviaguest
@russellpierce what about:
>  Functionality: Have the functional claims of the software been confirmed?
