+++ library.bib
@@ -3631,7 +3631,7 @@ appropriate for their analytic goals.},
   keywords = {clean_citation,paper_pass},
   title = {The Processing of Frequency and Duration},
   type = {Doctoral dissertation, Chemnitz University of Technology},
-  url = {http://monarch.qucosa.de/api/qucosa%3A19139/attachment/ATT-0/}
+  url = {https://nbn-resolving.org/urn:nbn:de:bsz:ch1-200900917}
 }
 
 @article{Winkler2015,
