+++ NEWS.md
@@ -1,4 +1,6 @@
-#
+# passt 0.1.2
+
+fixed url for Winkler dissertation
 
 # passt 0.1.1
 
