### 2020-02-05 #1900 @johannes-titz
@oliviaguest Thank you! There is a minor problem: one url does not resolve. I just fixed this and released version 0.1.2. The only change is the url in the bib file and the link in readme.md (see here: https://github.com/johannes-titz/passt/compare/v0.1.1...v0.1.2). Sorry for that. I believe the url resolved when I submitted. The updated link is now a persistent identifier (URN, comparable to doi).
