### 2020-02-05 #1900 @johannes-titz
@oliviaguest The bib file is auto-generated by zotero. It has a @preamble that is a command. If this causes problems, I can remove it.
