+++ library.bib
@@ -4121,6 +4121,3 @@ appropriate for their analytic goals.},
   title = {On the {{Nature}} and {{Analysis}} of {{Clustered Data}}},
   volume = {2}
 }
-
-@preamble{ "\ifdefined\DeclarePrefChars\DeclarePrefChars{'’-}\else\fi " }
-
