### 2020-02-05 #1900 @johannes-titz
@oliviaguest Sorry for the mess. My bib-file contains all papers I use. whedon checks them all, although only a small subset is used in the actual paper. The pdf looks fine to me. There are some references that do not have doi. I can also clean up the bib file.
