+++ NEWS.md
@@ -1,6 +1,8 @@
+#
+
 # passt 0.1.2
 
-fixed url for Winkler dissertation
+* fixed url for Winkler dissertation
 
 # passt 0.1.1
 
