### 2020-02-05 #1900 @johannes-titz
@oliviaguest Actually, there seem to be dois I was not aware of (e.g. for book chapters). If it is okay, I will sort this out. Sorry, again!
