+++ library.bib
@@ -83,6 +83,7 @@
   file = {/home/jt/1.paper/Sedlmeier - Improving statistical reasoning.pdf},
   keywords = {clean_citation,paper_pass},
   location = {{Mawah, NJ}},
+  doi = {10.4324/9781410601247},
   publisher = {{Erlbaum}},
   title = {Improving Statistical Reasoning: {{Theoretical}} Models and Practical Implications}
 }
