+++ library.bib
@@ -6,7 +6,7 @@
   keywords = {clean_citation,paper_pass},
   location = {{Oxford, England}},
   pages = {137-152},
-  doi = {https://doi.org/10.1093/acprof:oso/9780198508632.003.0009},
+  doi = {10.1093/acprof:oso/9780198508632.003.0009},
   publisher = {{Oxford University Press}},
   title = {Associative Learning and Frequency Judgments: {{The PASS}} Model}
 }
@@ -62,19 +62,6 @@
   howpublished = {Manuscript submitted for publication},
   keywords = {clean_citation,paper3_pass_t},
   title = {Simulating Asymmetries in Judgments of Frequency and Duration: {{The PASS}}-{{T}} Model}
-}
-
-@incollection{Sedlmeier2002,
-  author = {Sedlmeier, Peter and Betsch, Tilmann and Renkewitz, Frank},
-  booktitle = {Frequency Processing and Cognition},
-  date = {2002},
-  editor = {Sedlmeier, Peter and Betsch, Tilmann},
-  keywords = {clean_citation,paper_pass,paper3_pass_t},
-  location = {{Oxford, England}},
-  pages = {1-17},
-  doi = {https://doi.org/10.1093/acprof:oso/9780198508632.003.0001},
-  publisher = {{Oxford University Press}},
-  title = {Frequency Processing and Cognition: {{Introduction}} and Overview}
 }
 
 @book{Sedlmeier1999,
